refactor: merge frontmatter into template proc
